update-function-member-change-role-UI

+++ app/Http/Controllers/UserController.php
@@ -83,17 +83,14 @@ class UserController extends Controller
         $data = $request->validate(
             [
                 'name' => 'required|max:255',
-                'birthday' => 'required|max:255',
-                'favorite' => 'required|max:255',
-                'about' => 'required|max:255',
-                'cover' => 'required|image|mimes:jpg,png,jpeg,gif,svg|max:2048|dimensions:width=1110, height=300',
+                'birthday' => 'max:255',
+                'favorite' => 'max:255',
+                'about' => 'max:255',
+                'cover' => 'image|mimes:jpg,png,jpeg,gif,svg|max:2048|dimensions:width=1110, height=300',
             ],
             [
                 'name.required' => 'Phải có tên thành viên truyện!',
-                'birthday.required' => 'Ngày sinh không được bỏ trống!',
-                'favorite.required' => 'Sở thích không được bỏ trống!',
-                'about.required' => 'Phải có thông tin bản thân!',
-                'cover.required' => 'Phải có ảnh bìa!',
+                'cover.image' => 'Phải là file ảnh!',
                 'cover.dimensions' => 'Ảnh bìa phải có kích thước là 1110 x 300!',
             ]
         );
@@ -109,7 +106,9 @@ class UserController extends Controller
         if($get_image) {
             $path = 'uploads/user/'.$member->avatar;
             if(file_exists($path)) {
-                unlink($path);
+                if($path != 'uploads/user/default.png') {
+                    unlink($path);
+                }
             }
             $path = 'uploads/user/';
             $get_name_image = $get_image->getClientOriginalName();
@@ -123,7 +122,9 @@ class UserController extends Controller
         if($get_cover) {
             $path = 'uploads/user/'.$member->cover;
             if(file_exists($path)) {
-                unlink($path);
+                if($path != 'uploads/user/cover-default.jpg') {
+                    unlink($path);
+                }
             }
             $path = 'uploads/user/';
             $get_name_cover = $get_cover->getClientOriginalName();
@@ -139,5 +140,37 @@ class UserController extends Controller
         return redirect()->back()->with('status', 'Cập nhật thông tin thành công!');
     }
 
+
+    public function index()
+    {
+        $user = User::orderBy('id', 'DESC')->get();
+        return view('admin_cpanel.user.index')->with(compact('user'));
+    }
+
+    public function edit($id)
+    {
+        $user = User::find($id);
+        return view('admin_cpanel.user.edit')->with(compact('user'));
+    }
+
+    public function admin_update(Request $request, $id) {
+        $data = $request->validate(
+            [
+                'role' => 'required',
+            ],
+        );
+
+        $member = User::find($id);
+        $member->role = $data['role'];
+
+        $member->save();
+        return redirect()->back()->with('status', 'Cập nhật thành viên thành công!');
+    }
+
+    public function destroy($id)
+    {
+        $user = User::find($id)->destroy();
+        return redirect()->back()->with('status', 'Xóa thành viên thành công!');
+    }
 
 }
